refactored blocks
