+++ wagtail_plotly/blocks/__init__.py
@@ -1,0 +1,1 @@
+from .blocks import *    # noqa
